advanced-agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl-py>=2.8.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "weave>=0.51.53",
]
